// typescript/animationsXX/aniNoPerc/designNoPerc/preInitoscilate.d.ts
export default class PreInitOscilate {
    startTimeSec: number;
    endTimeSec: number;
    startValue: number;
    endValue: number;
    speed: number;
    constructor(startTimeSec: number, endTimeSec: number, pointOne?: number, pointTwo?: number, skipFrames?: number);
}
//# sourceMappingURL=preInitoscilate.d.ts.map